+++ jenkins-kk-tests/js/formControls/configForm.html
