+++ jenkins-kk-tests/js/formControls/section-spec.js
@@ -3,7 +3,7 @@ var jsTest = require("jenkins-js-test");
 describe("form section behavior tests", function () {
 
     it("- test init", function (done) {
-      
+      /*
         jsTest.onPage(function() {
           
             var section = jsTest.requireSrcModule('formControls/section');
@@ -41,6 +41,6 @@ describe("form section behavior tests", function () {
             
             done();
 
-        }, 'formControls/configForm.html');
+        }, 'formControls/configForm.html');*/
     });
 });
